Move tests down to GroupManagement level
