+++ src/components/messenger/list/index.test.tsx
@@ -17,7 +17,6 @@ import { RewardsState } from '../../../store/rewards';
 import { RewardsContainer } from '../../rewards-container';
 import { previewDisplayDate } from '../../../lib/chat/chat-message';
 import { SettingsMenu } from '../../settings-menu';
-import { AddMembersPanel } from './add-members-panel';
 import { GroupManagement } from './group-management';
 
 const mockSearchMyNetworksByName = jest.fn();
@@ -324,67 +323,16 @@ describe('messenger-list', () => {
     expect(enterFullScreenMessenger).toHaveBeenCalledOnce();
   });
 
-  describe('group management', () => {
-    const groupSubject = (props: Partial<Properties> = {}) => {
-      const wrapper = subject(props);
-      return wrapper.find(GroupManagement).dive();
-    };
+  it('renders GroupManagement if group management stage is NOT none', function () {
+    const wrapper = subject({ stage: Stage.None, groupManangemenetStage: GroupManagementStage.StartAddMemberToRoom });
 
-    it('renders AddMembersPanel', function () {
-      let wrapper = groupSubject({
-        stage: Stage.None,
-        groupManangemenetStage: GroupManagementStage.StartAddMemberToRoom,
-      });
+    expect(wrapper).toHaveElement(GroupManagement);
+  });
 
-      expect(wrapper).not.toHaveElement(ConversationListPanel);
-      expect(wrapper).not.toHaveElement(CreateConversationPanel);
-      expect(wrapper).not.toHaveElement(StartGroupPanel);
-      expect(wrapper).not.toHaveElement(GroupDetailsPanel);
-      expect(wrapper).toHaveElement(AddMembersPanel);
-    });
+  it('does not render GroupManagement if group management stage is none', function () {
+    const wrapper = subject({ stage: Stage.None, groupManangemenetStage: GroupManagementStage.None });
 
-    it('does not render AddMembersPanel if group management stage is none', function () {
-      const wrapper = subject({ stage: Stage.None, groupManangemenetStage: GroupManagementStage.None });
-
-      expect(wrapper).not.toHaveElement(GroupManagement);
-    });
-
-    it('moves back from AddMembersPanel', async function () {
-      const backGroupManagement = jest.fn();
-      const wrapper = groupSubject({
-        stage: Stage.None,
-        groupManangemenetStage: GroupManagementStage.StartAddMemberToRoom,
-        backGroupManagement,
-      });
-
-      await wrapper.find(AddMembersPanel).simulate('back');
-
-      expect(backGroupManagement).toHaveBeenCalledOnce();
-    });
-
-    it('searches for citizens when adding new members', async function () {
-      when(mockSearchMyNetworksByName)
-        .calledWith('jac')
-        .mockResolvedValue([{ id: 'user-id', profileImage: 'image-url' }]);
-      const wrapper = groupSubject({
-        stage: Stage.None,
-        groupManangemenetStage: GroupManagementStage.StartAddMemberToRoom,
-      });
-
-      const searchResults = await wrapper.find(AddMembersPanel).prop('searchUsers')('jac');
-
-      expect(searchResults).toStrictEqual([{ id: 'user-id', image: 'image-url', profileImage: 'image-url' }]);
-    });
-
-    it('sets AddMembersPanel to Submitting while data is loading', async function () {
-      const wrapper = groupSubject({
-        stage: Stage.None,
-        groupManangemenetStage: GroupManagementStage.StartAddMemberToRoom,
-        isFetchingExistingConversations: true,
-      });
-
-      expect(wrapper.find(AddMembersPanel).prop('isSubmitting')).toBeTrue();
-    });
+    expect(wrapper).not.toHaveElement(GroupManagement);
   });
 
   describe('mapState', () => {
